src/requirements.txt
# Core
Django>=4.2.0
django-ninja>=1.0.1
django-cors-headers>=4.3.0
social-auth-app-django>=5.4.0
python-dotenv>=1.0.0
django-ninja-jwt>=5.2.5
djangorestframework>=3.14.0
django-ninja-extra>=0.19.8

# Data Processing & Analysis
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.1.0

# Caching & Performance
django-redis>=5.4.0
django-cacheops>=7.0.0

# Export Formats
openpyxl>=3.1.0  # For Excel export support

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
factory-boy>=3.3.0

# Development
black>=23.10.0  # Code formatting
isort>=5.12.0   # Import sorting
flake8>=6.1.0   # Code linting 